Implemented hunger & score
frog gets smaller if it doesn't get the flies before they reach the other side of the screen, and it dies if it misses 3 flies in a row. eating a fly replenishes hunger and the frog grows big again. amount of flies eaten is counted and displayed on the screen.