added tile spawner
now smaller bricks are spawned in a grid with a for each loop. the grid spawns right horizontally until it reaches a calculated index (right side of the grid), then it moves up a row and back to the left keeps spawning. right now the grid array is massive with 1 element for each tile, which will be changed in the future.