random brick pattern
bricks spawn with a random health from 0 to 3, and if their health is 0 they are spliced from the array. This creates a random pattern on the grid. There is a bug where brick colors flicker when the ball hits a brick near them, I think it is related to the corners bug from collisions.